refactor: global 패키지 분리
